src/auth/jwt_utils.py
from datetime import timedelta, timezone, datetime
from typing import Annotated

import jwt
from fastapi import Depends, HTTPException
from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
from jwt import InvalidTokenError
from starlette import status

from src.auth.schemas import TokenSchema
from src.auth.service import AuthService
from src.auth.pwd_utils import verify_password
from src.config import settings

oauth2_scheme = OAuth2PasswordBearer(tokenUrl="auth/login")

# to get a string like this run:
# openssl rand -hex 32
SECRET_KEY = settings.SECRET_KEY
ALGORITHM = settings.ALGORITHM
ACCESS_TOKEN_EXPIRE_MINUTES = 30


def create_access_token(data: dict, expires_delta: timedelta | None = None):
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.now(timezone.utc) + expires_delta
    else:
        expire = datetime.now(timezone.utc) + timedelta(minutes=15)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt


def get_current_token_payload(
    token: str,
):
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
    except InvalidTokenError as e:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail=f"Invalid token error: {e}",
        )
    return payload


async def get_current_user(token: Annotated[str, Depends(oauth2_scheme)]):
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        username: str = payload.get("sub")
        if username is None:
            raise credentials_exception
    except InvalidTokenError:
        raise credentials_exception
    user = await AuthService.get_user_by_username(username)
    if user is None:
        raise credentials_exception
    return user


async def get_access_token(form_data: OAuth2PasswordRequestForm) -> TokenSchema | None:
    username = form_data.username
    user = await AuthService.get_user_by_username(username)
    data_dict = {"sub": username}
    access_token = create_access_token(data_dict)

    if user is None:
        return None
    if not verify_password(form_data.password, user.hashed_password):
        return None
    return TokenSchema(access_token=access_token, token_type="bearer")
