requirements.txt
aiosqlite==0.20.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
bcrypt==4.2.1
black==24.10.0
build==1.2.2.post1
CacheControl==0.14.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
crashtest==0.4.1
cryptography==44.0.0
distlib==0.3.9
dnspython==2.7.0
dulwich==0.22.7
email_validator==2.2.0
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.17.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
Jinja2==3.1.5
keyring==25.6.0
Mako==1.3.8
MarkupSafe==3.0.2
more-itertools==10.6.0
msgpack==1.1.0
mypy-extensions==1.0.0
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.5.0
poetry==2.0.1
poetry-core==2.0.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
pyproject_hooks==1.2.0
pytest==8.3.4
pytest-asyncio==0.25.2
python-dotenv==1.0.1
python-multipart==0.0.20
RapidFuzz==3.11.0
requests==2.32.3
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.4
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.37
starlette==0.41.3
tomlkit==0.13.2
trove-classifiers==2025.1.15.22
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.1
WTForms==3.1.2
